Initial Commit
cmdlet to publish the current local version of the azure function to remote. Please note, I have found that the syncing of triggers can fail if the function is "stopped" be sure to "start" the function before building to remote.

Also, this will delete the existing remote build and replace with the local version. Be sure the local version is complete.